# generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Gender": {
				"type": "string",
				"enum": [
					"male",
					"female",
					"other"
				]
			},
			"FitnessGoal": {
				"type": "string",
				"enum": [
					"general_fitness",
					"fat_loss",
					"endurance",
					"muscle_gain",
					"strength",
					"mobility_flexibility",
					"balance",
					"recovery"
				]
			},
			"FitnessLevel": {
				"type": "string",
				"enum": [
					"beginner",
					"intermediate",
					"advanced"
				]
			},
			"PhysicalLimitation": {
				"type": "string",
				"enum": [
					"knee_pain",
					"shoulder_pain",
					"lower_back_pain",
					"neck_pain",
					"hip_pain",
					"ankle_instability",
					"wrist_pain",
					"elbow_pain",
					"arthritis",
					"osteoporosis",
					"sciatica",
					"limited_range_of_motion",
					"post_surgery_recovery",
					"balance_issues",
					"chronic_fatigue",
					"breathing_issues"
				]
			},
			"WorkoutEnvironment": {
				"type": "string",
				"enum": [
					"home_gym",
					"commercial_gym",
					"bodyweight_only"
				]
			},
			"AvailableEquipment": {
				"type": "string",
				"enum": [
					"barbells",
					"bench",
					"incline_decline_bench",
					"pull_up_bar",
					"bike",
					"medicine_balls",
					"plyo_box",
					"rings",
					"resistance_bands",
					"stability_ball",
					"dumbbells",
					"kettlebells",
					"squat_rack",
					"dip_bar",
					"rowing_machine",
					"slam_balls",
					"cable_machine",
					"jump_rope",
					"foam_roller"
				]
			},
			"PreferredStyles": {
				"type": "string",
				"enum": [
					"strength",
					"balance",
					"hiit",
					"cardio",
					"rehab",
					"crossfit",
					"functional",
					"pilates",
					"yoga",
					"mobility"
				]
			},
			"PreferredDay": {
				"type": "string",
				"enum": [
					"monday",
					"tuesday",
					"wednesday",
					"thursday",
					"friday",
					"saturday",
					"sunday"
				]
			},
			"IntensityLevel": {
				"type": "string",
				"enum": [
					"low",
					"moderate",
					"high"
				]
			},
			"Profile": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"goals": {
						"items": {
							"$ref": "#/components/schemas/FitnessGoal"
						},
						"type": "array"
					},
					"fitnessLevel": {
						"$ref": "#/components/schemas/FitnessLevel"
					},
					"limitations": {
						"items": {
							"$ref": "#/components/schemas/PhysicalLimitation"
						},
						"type": "array"
					},
					"medicalNotes": {
						"type": "string"
					},
					"environment": {
						"$ref": "#/components/schemas/WorkoutEnvironment"
					},
					"equipment": {
						"items": {
							"$ref": "#/components/schemas/AvailableEquipment"
						},
						"type": "array"
					},
					"otherEquipment": {
						"type": "string"
					},
					"preferredStyles": {
						"items": {
							"$ref": "#/components/schemas/PreferredStyles"
						},
						"type": "array"
					},
					"availableDays": {
						"items": {
							"$ref": "#/components/schemas/PreferredDay"
						},
						"type": "array"
					},
					"workoutDuration": {
						"type": "number",
						"format": "double"
					},
					"intensityLevel": {
						"$ref": "#/components/schemas/IntensityLevel"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"needsOnboarding": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"emailVerified",
					"createdAt",
					"needsOnboarding"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"profile": {
						"$ref": "#/components/schemas/Profile"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"needsOnboarding": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"profile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Profile_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"goals": {
						"items": {
							"$ref": "#/components/schemas/FitnessGoal"
						},
						"type": "array"
					},
					"fitnessLevel": {
						"$ref": "#/components/schemas/FitnessLevel"
					},
					"limitations": {
						"items": {
							"$ref": "#/components/schemas/PhysicalLimitation"
						},
						"type": "array"
					},
					"medicalNotes": {
						"type": "string"
					},
					"environment": {
						"$ref": "#/components/schemas/WorkoutEnvironment"
					},
					"equipment": {
						"items": {
							"$ref": "#/components/schemas/AvailableEquipment"
						},
						"type": "array"
					},
					"otherEquipment": {
						"type": "string"
					},
					"preferredStyles": {
						"items": {
							"$ref": "#/components/schemas/PreferredStyles"
						},
						"type": "array"
					},
					"availableDays": {
						"items": {
							"$ref": "#/components/schemas/PreferredDay"
						},
						"type": "array"
					},
					"workoutDuration": {
						"type": "number",
						"format": "double"
					},
					"intensityLevel": {
						"$ref": "#/components/schemas/IntensityLevel"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "mastersfit-server",
		"version": "1.0.0",
		"description": "RESTful API server for Masters Fit fitness tracking application",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/profile/{userId}": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get the user's profile",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/profile": {
			"post": {
				"operationId": "CreateProfile",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create or update the user's profile",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Profile_"
							}
						}
					}
				}
			}
		},
		"/profile/{id}": {
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update the user's profile",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Profile_"
							}
						}
					}
				}
			}
		},
		"/profile/user/{userId}": {
			"put": {
				"operationId": "UpdateProfileByUserId",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update the user's profile by userId",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Profile_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}